[CodeClean] typecast of size
Code clean to fix svace issue, typecast of size and index value.

+++ ext/nnstreamer/tensor_filter/tensor_filter_movidius_ncsdk2.c
@@ -190,7 +190,7 @@ _mvncsdk2_open (const GstTensorFilterProperties * prop, void **private_data)
   }
 
   /* Warning: conversion unsigned long to unsigned int */
-  len_model_file = g_mapped_file_get_length (file_model);
+  len_model_file = (guint32) g_mapped_file_get_length (file_model);
   buf_model_file = (void *) g_mapped_file_get_contents (file_model);
 
   /* Actually, send the model to the device */
